+++ perun/logic/config.py
@@ -49,7 +49,7 @@ class Config(object):
                 'editor': 'vim'
             },
             'format': {
-                'status': '%type% > %origin%',
+                'status': '%type% > %source%',
                 'log': '%checksum:6%: %desc%'
             }
         }
@@ -163,7 +163,7 @@ general:
     paging: only-log
 
 format:
-    status: "\u2503 %type% \u2503 %collector%  \u2503 (%time%) \u2503 %origin% \u2503"
+    status: "\u2503 %type% \u2503 %collector%  \u2503 (%time%) \u2503 %source% \u2503"
     shortlog: "%checksum:6% (%stats%) %desc%"
     output_profile_template: "%collector%-%cmd%-%args%-%workload%-%date%"
     output_show_template: "%collector%-%cmd%-%args%-%workload%-%date%"