+++ tests/logic/commands/test_cli.py
@@ -247,7 +247,7 @@ def test_status_correct(pcs_full):
     assert short_result.exit_code == 0
     assert len(short_result.output.split("\n")) == 4
 
-    short_result = runner.invoke(cli.status, ['--short', '--sort-by', 'origin'])
+    short_result = runner.invoke(cli.status, ['--short', '--sort-by', 'source'])
     assert short_result.exit_code == 0