+++ tests/logic/commands/test_profile.py
@@ -1,9 +1,34 @@
 import pytest
+import git
 
+import perun.logic.commands as commands
 import perun.logic.config as config
 import perun.profile.factory as factory
 
 __author__ = 'Tomas Fiedor'
+
+
+def test_loading(helpers, pcs_full, valid_profile_pool):
+    """Test new feature of loading the profile straight out of profile info
+
+    Expecting correct behaviour
+    """
+    helpers.populate_repo_with_untracked_profiles(pcs_full.path, valid_profile_pool)
+    untracked = commands.get_untracked_profiles(pcs_full)
+    assert len(untracked) != 0
+
+    first_untracked = untracked[0].load()
+    assert isinstance(first_untracked, dict)
+    assert 'header' in first_untracked.keys()
+
+    git_repo = git.Repo(pcs_full.vcs_path)
+    head = str(git_repo.head.commit)
+
+    minor_version_profiles = commands.get_minor_version_profiles(pcs_full, head)
+    assert len(minor_version_profiles) != 0
+    first_indexed = minor_version_profiles[0].load()
+    assert isinstance(first_indexed, dict)
+    assert 'header' in first_indexed.keys()
 
 
 def test_name_generation(capsys):